fix email not displaying properly on settings screen
