+++ apps/android-app/lib/app/app_router.dart
@@ -10,7 +10,7 @@ import '../features/login/presentation/auth_page.dart';
 import '../features/settings/presentation/screens/profile_page.dart';
 import '../features/path/presentation/screens/workouts_page.dart';
 import '../features/path/presentation/screens/workout_page.dart';
-import '../features/path/domain/entities/WorkoutEntity.dart';
+import '../features/path/domain/entities/workout_entity.dart';
 
 
 part 'app_router.gr.dart';
