+++ apps/android-app/lib/app/data/services/network_service.dart
@@ -45,7 +45,7 @@ class DioClient {
   DioClient._internal()
     : _dio = Dio(
         BaseOptions(
-          baseUrl: 'http://10.0.2.2:8080',
+          baseUrl: AppVariables.baseUrl,
           connectTimeout: const Duration(seconds: 10),
           receiveTimeout: const Duration(seconds: 10),
           contentType: 'application/json',
