+++ apps/android-app/lib/features/path/data/repositories/workout_path_network_repository.dart
@@ -1,5 +1,5 @@
 import 'package:android_app/app/data/services/network_service.dart';
-import 'package:android_app/features/path/domain/entities/WorkoutEntity.dart';
+import 'package:android_app/features/path/domain/entities/workout_entity.dart';
 import 'package:android_app/features/path/domain/repositories/workout_path_repository.dart';
 import 'package:flutter/foundation.dart';
 class WorkoutPathNetworkRepository extends WorkoutPathRepository {
