+++ apps/android-app/lib/features/path/domain/repositories/workout_path_repository.dart
@@ -1,4 +1,4 @@
-import 'package:android_app/features/path/domain/entities/WorkoutEntity.dart';
+import 'package:android_app/features/path/domain/entities/workout_entity.dart';
 
 abstract class WorkoutPathRepository {
   Future<List<WorkoutEntity>> fetchWorkouts();
