+++ apps/android-app/pubspec.yaml
@@ -38,6 +38,7 @@ dependencies:
   dio: any
   dio_smart_retry: any
   flutter_dotenv: any
+  http: any
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
